refactor(observability): 归位 Conversation 与调试能力
