+++ src/observability/conversation/index-process.ts
@@ -4,8 +4,8 @@ import {
   extendCodexRolloutIndex,
   isReusableCodexRolloutIndex,
   type CodexRolloutIndex,
-} from './trace/adapters/codex/conversation-index.js';
-import { writeJsonFileAtomic } from '../shared/atomic-json.js';
+} from '../trace/adapters/codex/conversation-index.js';
+import { writeJsonFileAtomic } from '../../shared/atomic-json.js';
 
 const [sourcePath, sourceThreadId, cachePath] = process.argv.slice(2);
