+++ src/observability/experience.ts
@@ -84,7 +84,7 @@ import {
   type SkillSegment,
 } from './trace/segmentation.js';
 import { createTraceSessionIndex, traceSessionRefIdentity } from './trace/session-index.js';
-import { reconstructExperienceTurns } from './turn-index.js';
+import { reconstructExperienceTurns } from './conversation/turn-index.js';
 import { isToolResultFailureText } from '../executors/tool-call-status.js';
 import { hasAssistantDeliverableArtifactText, hasUserHardRuleText, isSyntheticUserMessageText, isUserInteractionMetricText, USER_INTERRUPTION_RE } from './text-signals.js';
 import {
