+++ test/studio/http/conversation-activity-server.test.ts
@@ -3,7 +3,7 @@ import { mkdirSync, mkdtempSync, rmSync } from 'node:fs';
 import { tmpdir } from 'node:os';
 import { join } from 'node:path';
 import { afterAll, beforeAll, describe, it } from 'vitest';
-import type { ConversationCatalog } from '../../../src/observability/conversation-catalog.js';
+import type { ConversationCatalog } from '../../../src/observability/conversation/catalog.js';
 import { createReportServer } from '../../../src/studio/http/report-server.js';
 import type { ConversationListItem } from '../../../src/observability/view-models/conversation.js';
